requirements.txt
asgiref==3.9.1
Django==3.2
djangorestframework==3.12.4
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.1
